move stories out to separate directory
